requirements.txt
attrs >= 19.3.0